requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-crispy-forms==1.14.0
django-filter==21.1
django-login-required-middleware==0.7.0
django-widget-tweaks==1.4.12
Pillow==9.0.1
python-barcode==0.13.1
sqlparse==0.4.2
